Started some refactoring
The new core package is supposed to hold all support code, that can be
re-used in other projects. The ultimate goal is it, to extract it into
it's own project and with it stuff like the Lost Evolution code.

Removed the COLLADA related code. It didn't work out.

Added some JavaDocs to some classes.